+++ packages/common/CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## 7.0.1 2023-05-25
-
 ### Fixed
 
 - Cost calculation for `deployModule` transaction.
+
+### Added
+
+- A `parseWallet` function to parse wallet export files
 
 ## 7.0.0 2023-05-15